# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Quản Lý Cân Xe API",
    "version": "1.0.0",
    "description": "API tài liệu cho hệ thống quản lý cân xe.\n\nHệ thống này cung cấp API để quản lý việc cân xe, thông tin khách hàng, \nhàng hóa và các chức năng liên quan đến quản lý phiếu cân.\n",
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Your Name",
      "url": "https://yourwebsite.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Soxe": {
        "type": "object",
        "required": [
          "soxe"
        ],
        "properties": {
          "id": {
            "type": "number",
            "description": "ID tự động tạo của số xe"
          },
          "soxe": {
            "type": "string",
            "description": "Biển số xe"
          },
          "trongluong": {
            "type": "number",
            "description": "Trọng lượng xe (kg)"
          }
        },
        "example": {
          "id": 1,
          "soxe": "51F-12345",
          "trongluong": 15000
        }
      },
      "SoxeCreate": {
        "type": "object",
        "required": [
          "soxe"
        ],
        "properties": {
          "soxe": {
            "type": "string",
            "description": "Biển số xe"
          },
          "trongluong": {
            "type": "number",
            "description": "Trọng lượng xe (kg)"
          }
        },
        "example": {
          "soxe": "51F-12345",
          "trongluong": 15000
        }
      },
      "SoxeUpdate": {
        "type": "object",
        "properties": {
          "soxe": {
            "type": "string",
            "description": "Biển số xe"
          },
          "trongluong": {
            "type": "number",
            "description": "Trọng lượng xe (kg)"
          }
        },
        "example": {
          "trongluong": 16000
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/soxe": {
      "get": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Lấy danh sách số xe",
        "description": "Lấy toàn bộ danh sách số xe trong hệ thống",
        "operationId": "getAllSoxes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách số xe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Danh sách số xe"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Soxe"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Tạo số xe mới",
        "description": "Tạo một số xe mới trong hệ thống",
        "operationId": "createSoxe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoxeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo số xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo số xe thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Soxe"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/soxe/{id}": {
      "get": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Lấy thông tin số xe theo ID",
        "description": "Lấy chi tiết một số xe dựa trên ID",
        "operationId": "getSoxeById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết số xe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết số xe"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Soxe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Cập nhật số xe",
        "description": "Cập nhật thông tin số xe dựa trên ID",
        "operationId": "updateSoxe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoxeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật số xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật số xe thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Soxe"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Xóa số xe",
        "description": "Xóa một số xe dựa trên ID",
        "operationId": "deleteSoxe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa số xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa số xe thành công"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/soxe/number/{soxe}": {
      "get": {
        "tags": [
          "Số Xe"
        ],
        "summary": "Lấy thông tin số xe theo biển số",
        "description": "Lấy chi tiết một số xe dựa trên biển số",
        "operationId": "getSoxeBySoxe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "soxe",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Biển số xe"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết số xe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết số xe"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Soxe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  }
}